Update Unit Tests
Split them out to different files.
Fixup some things and bring up numbers a little.
